requirements.txt
setuptools>=65.5.0
wheel>=0.38.0
flask==2.0.1
flask-cors==3.0.10
requests==2.31.0
python-dotenv==0.19.0
flask-limiter==3.5.0
pydub==0.25.1
numpy>=1.24.0
openai>=1.10.0
gunicorn==21.2.0
google-auth==2.29.0
google-auth-oauthlib==0.4.6
google-auth-httplib2==0.1.0
google-api-python-client==2.31.0
twilio==8.13.0
pyyaml==6.0
crewai
langchain
langchain-openai
duckduckgo-search==4.1.1
python-docx==1.1.0
markdown2==2.4.10 